Moved BackupContainerAzureBlobStore into its own files
